Home Route Functional.
