--- Auth/.env
@@ -1,2 +1,0 @@
-PORT=9000
-MONGODB_URL=mongodb+srv://admin:admin@cluster0.3snih.mongodb.net/user?retryWrites=true&w=majority
