--- Auth/controllers/eventController.go
@@ -1,103 +1,0 @@
-package controllers
-
-import (
-	"context"
-	"fmt"
-	"net/http"
-	"time"
-
-	"github.com/SimranBhagwandasani/EventPlanner/database"
-	"github.com/SimranBhagwandasani/EventPlanner/models"
-	"github.com/gin-gonic/gin"
-	"github.com/go-playground/validator/v10"
-
-	"go.mongodb.org/mongo-driver/bson/primitive"
-	"go.mongodb.org/mongo-driver/mongo"
-
-	"log"
-	"strconv"
-
-	"go.mongodb.org/mongo-driver/bson"
-)
-
-var eventCollection *mongo.Collection = database.OpenCollection(database.Client, "events")
-var validateEvent = validator.New()
-
-func CreateEvent() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		c.Header("Content-Type", "application/x-www-form-urlencoded")
-		c.Header("Access-Control-Allow-Origin","*")
-		c.Header("Access-Control-Allow-Methods","POST")
-		c.Header("Access-Control-Allow-Headers","Content-Type")
-		fmt.Println("called Events Route")
-		var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
-		var event models.Event
-
-		if err := c.BindJSON(&event); err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error_aa": err.Error()})
-			return
-		}
-
-		validationErr := validateEvent.Struct(event)
-		if validationErr != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error_vv": validationErr.Error()})
-			return
-		}
-		event.Created_at, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
-		event.Updated_at, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
-
-		event.ID = primitive.NewObjectID()
-		event.Event_id = event.ID.Hex()
-
-		resultInsertionNumber, insertErr := eventCollection.InsertOne(ctx, event)
-		if insertErr != nil {
-			fmt.Println("Inserted")
-			msg := fmt.Sprintf("User item was not created")
-			c.JSON(http.StatusInternalServerError, gin.H{"error_message": msg})
-			return
-		}
-
-		c.JSON(http.StatusOK, resultInsertionNumber)
-		defer cancel()
-		return
-	}
-}
-func GetEvents() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		var ctx, cancel = context.WithTimeout(context.Background(), 100*time.Second)
-
-		recordPerPage, err := strconv.Atoi(c.Query("recordPerPage"))
-		if err != nil || recordPerPage < 1 {
-			recordPerPage = 10
-		}
-		page, err1 := strconv.Atoi(c.Query("page"))
-		if err1 != nil || page < 1 {
-			page = 1
-		}
-
-		startIndex := (page - 1) * recordPerPage
-		startIndex, err = strconv.Atoi(c.Query("startIndex"))
-
-		matchStage := bson.D{{"$match", bson.D{{}}}}
-		groupStage := bson.D{{"$group", bson.D{
-			{"_id", bson.D{{"_id", "null"}}},
-			{"total_count", bson.D{{"$sum", 1}}},
-			{"data", bson.D{{"$push", "$$ROOT"}}}}}}
-		projectStage := bson.D{
-			{"$project", bson.D{
-				{"_id", 0},
-				{"total_count", 1},
-				{"user_items", bson.D{{"$slice", []interface{}{"$data", startIndex, recordPerPage}}}}}}}
-		result, err := eventCollection.Aggregate(ctx, mongo.Pipeline{
-			matchStage, groupStage, projectStage})
-		defer cancel()
-		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "error occured while listing user items"})
-		}
-		var allusers []bson.M
-		if err = result.All(ctx, &allusers); err != nil {
-			log.Fatal(err)
-		}
-		c.JSON(http.StatusOK, allusers[0])
-	}
-}
