+++ Server/database/databaseConnection.go
@@ -1,0 +1,44 @@
+package database
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"os"
+	"time"
+
+	"github.com/joho/godotenv"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+func DBinstance() *mongo.Client {
+	err := godotenv.Load(".env")
+	if err != nil {
+		log.Fatal("Error Loading .env file")
+	}
+	MongoDb := os.Getenv("MONGODB_URL")
+	client, err := mongo.NewClient(options.Client().ApplyURI(MongoDb))
+	if err != nil {
+		log.Fatal(err)
+	} else {
+		log.Printf("Mongo cluster added..")
+	}
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+
+	err = client.Connect(ctx)
+	if err != nil {
+		log.Fatal(err)
+	}
+	fmt.Println("Connected to MongoDB!")
+	return client
+
+}
+
+var Client *mongo.Client = DBinstance()
+
+func OpenCollection(client *mongo.Client, collectionName string) *mongo.Collection {
+	var collection *mongo.Collection = client.Database("cluster0").Collection(collectionName)
+	return collection
+}
