+++ Server/helpers/authHelper.go
