+++ Server/main.go
@@ -3,33 +3,29 @@ package main
 import (
 	"fmt"
 	"os"
+
+	routes "github.com/SimranBhagwandasani/EventPlanner/routes"
 	"github.com/gin-gonic/gin"
-	"github.com/SimranBhagwandasani/EventPlanner/database"
-	"github.com/SimranBhagwandasani/EventPlanner/routes"
-	"github.com/SimranBhagwandasani/EventPlanner/middleware"
-	"go.mongodb.org/mongo-driver/mongo"
 )
-var eventCollection *mongo.collection = database.OpenCollection(database.Client,"food")
+
 func main() {
-	fmt.Print("Welcome to the Event planner.")
 	port := os.Getenv("PORT")
 
 	if port == "" {
 		port = "8000"
 	}
 
+	fmt.Println("Server running on port:", port)
 	router := gin.New()
 	router.Use(gin.Logger())
+
+	routes.AuthRoutes(router)
 	routes.UserRoutes(router)
-	router.Use(middleware.Authentication())
-
 	routes.EventRoutes(router)
-	routes.CatalogRoutes(router)
-	routes.TableRoutes(router)
-	routes.OrderRoutes(router)
-	routes.OrderItemRoutes(router)
-	routes.InvoiceRoutes(router)
+	routes.TicketRoutes(router)
+	routes.HomeRoutes(router)
 
-	router.Run(":"+port)
+	router.Run(":" + port)
 
+	fmt.Println("Auth testing.")
 }
