+++ Server/routes/userRouter.go
@@ -1,13 +1,28 @@
 package routes
 
-import(
-	"github.com/gin-gonic/gin"
+import (
 	controller "github.com/SimranBhagwandasani/EventPlanner/controllers"
+	"github.com/SimranBhagwandasani/EventPlanner/middleware"
+
+	"github.com/gin-gonic/gin"
 )
 
-func UserRoutes(incomingRoutes *gin.Engine){
-	incomingRoutes.GET("/users",controller.GetUsers())
-	incomingRoutes.GET("/users/:user_id",controller.GetUser())
-	incomingRoutes.POST("/users/signup",controller.SignUp())
-	incomingRoutes.POST("/users/login",controller.Login())
+func UserRoutes(incomingRoutes *gin.Engine) {
+	incomingRoutes.Use(middleware.Authenticate())
+	incomingRoutes.GET("/users", controller.GetUsers())
+	incomingRoutes.GET("/users/:user_id", controller.GetUser())
+}
+func HomeRoutes(incomingRoutes *gin.Engine) {
+	incomingRoutes.GET("/home", controller.GetEvents())
+}
+
+func EventRoutes(incomingRoutes *gin.Engine) {
+	incomingRoutes.Use(middleware.Authenticate())
+	incomingRoutes.POST("/event", controller.CreateEvent())
+	incomingRoutes.GET("/event", controller.GetEvents())
+}
+
+func TicketRoutes(incomingRoutes *gin.Engine) {
+	incomingRoutes.Use(middleware.Authenticate())
+	incomingRoutes.POST("/ticket", controller.PurchaseTicket())
 }
